third_party/blink/web_tests/jquery/attributes-expected.txt

ALERT: 339 tests of 339 passed, 0 failed.
